refactor: not sea bed but nor ocean floor & pathes
