+++ src/components/models/OceanFloor.vue
@@ -1,22 +1,22 @@
 <script setup lang="ts">
-import { getNode, SeaBedScene } from '@/assets/models/SeaBed/SeaBed.gltf.js'
+import { getNode, OceanFloorScene } from '@/assets/models/OceanFloor/OceanFloor.gltf.js'
 import { addToGroup, overrideFogShader } from '@/utils/threeHelper.js'
 import { Vector3, type WebGLProgramParametersWithUniforms } from 'three'
 import { shallowRef, watch } from 'vue'
 
 const groupWrapperRef = shallowRef()
 
-const model = await getNode(SeaBedScene.zligislandlvl0_seabed001)
-const seabedPlaneMaterial = (() => { // do not create a new material but use the existing one by cloning it
+const model = await getNode(OceanFloorScene.ocean_floor)
+const material = (() => { // do not create a new material but use the existing one by cloning it
   if (Array.isArray(model.material)) {
     return model.material[0].clone()
   }
   return model.material.clone()
 })()
-seabedPlaneMaterial.onBeforeCompile = function (shader: WebGLProgramParametersWithUniforms) {
+material.onBeforeCompile = function (shader: WebGLProgramParametersWithUniforms) {
   overrideFogShader(shader, new Vector3(0, 0, 0), 0)
 }
-model.material = seabedPlaneMaterial
+model.material = material
 
 const { stop } = watch(groupWrapperRef, (newValue) => {
   if (!newValue) {
