+++ src/components/models/SkyDome.vue
